Move CompactLevel0Phase1Result and metrics structs to compaction.rs
